refactor(security): split ssrf.rs into module to clear 800-LOC cap
Pure code-move; public path crate::security::ssrf::* unchanged via
re-exports. No behavior change (MAX_REDIRECT_HOPS=5, per-hop
revalidation, fail-closed all identical). Closes the sole DoD yellow
on PR #339.

Local-Tested: cargo fmt+clippy+test green @ 54eb60b9604840a6510478af1147abbfd1cf83be
